+++ submitr/submission_uploads.py
@@ -201,14 +201,11 @@ def get_submission_object_upload_files(submission_item: dict, portal: Portal) ->
 
 
 def upload_files(files: List[FileForUpload], portal: Portal) -> None:
-    if isinstance(files, list) and files:
-        files = [file for file in files if not file.ignore]
-        if not files:
-            PRINT("No files to upload.")
-            return
-        if yes_or_no(f"Ready to actually upload ({len(files)}) file{'s' if len(files) != 1 else ''}. Upload now?"):
-            for file in [file for file in files if not file.ignore]:
-                upload_file(file, portal=portal)
+    if not isinstance(files, list) or not files or not (files := [file for file in files if not file.ignore]):
+        PRINT("No files to upload.")
+    if yes_or_no(f"Ready to actually upload ({len(files)}) file{'s' if len(files) != 1 else ''}. Upload now?"):
+        for file in files:
+            upload_file(file, portal=portal)
 
 
 def upload_file(file: FileForUpload, portal: Portal) -> None: